+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+* Fixed: Exception when a command was sent to the bot via DM.
 * Added: Reminder system via the `!reminders` command.
 * Added: If a subcommand is missing, help will be displayed for that particular command.
 * Added: The `!help` command now shows commands with available help.